Refactor for game level data support

+++ source/game_objects.cpp
@@ -1,18 +1,18 @@
 #include "game_objects.h"
 
-void game_object::Update(double seconds)
+void game_object::Update(float seconds)
 {
   xPos += xVelocity;
   yPos += yVelocity;
 }
 
-void game_object::Accelerate(double increase)
+void game_object::Accelerate(float increase)
 {
   yVelocity -= increase * cos(DEGTORAD(angle));
   xVelocity += increase * sin(DEGTORAD(angle));
 }
 
-void bullet::Update(double timespanSeconds)
+void bullet::Update(float timespanSeconds)
 {
   gameObject.Update(timespanSeconds);
   lifespanSeconds -= timespanSeconds;
