+++ source/perf_data.cpp
@@ -4,6 +4,6 @@ perf_data::perf_data(const LARGE_INTEGER& perfFrequency, const LARGE_INTEGER& in
 {
   totalTicks = ticks.QuadPart - initialTicks.QuadPart;
   frameTicks = ticks.QuadPart - previousTicks.QuadPart;
-  frameTimeSeconds = static_cast<double>(frameTicks) / static_cast<double>(perfFrequency.QuadPart);
+  frameTimeSeconds = static_cast<float>(frameTicks) / static_cast<float>(perfFrequency.QuadPart);
   fps = frameTicks ? perfFrequency.QuadPart / frameTicks : 0;
 }
