+++ source/perf_data.h
@@ -7,6 +7,6 @@ struct perf_data
 
   int64_t totalTicks;
   int64_t frameTicks;
-  double frameTimeSeconds;
+  float frameTimeSeconds;
   int64_t fps;
 };
